% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgb-package.R
\docType{package}
\name{kgb-package}
\alias{kgb}
\alias{kgb-package}
\title{kgb: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/kgb}
  \item Report bugs at \url{https://github.com/frbcesab/kgb/issues}
}

}
\author{
\strong{Maintainer}: Camille Coux \email{camille.coux@orange.fr} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
